+++ dev4press/core/features/Load.php
@@ -358,10 +358,12 @@ abstract class Load {
 	}
 
 	public function get_for_edit( string $feature, string $name, $default = null ) {
-		$real = $feature . '__' . $name;
+		$real   = $feature . '__' . $name;
 
 		if ( $this->network_mode() && ! is_network_admin() ) {
-			return $this->b()->get( $real, 'features', $default );
+			if ( $this->allow_blog_override( $feature ) ) {
+				return $this->b()->get( $real, 'features', $default );
+			}
 		}
 
 		return $this->s()->get( $real, 'features', $default );